Improved python interface. Added scale method for basic set classes.

+++ include/declarations.h
@@ -33,7 +33,7 @@
 #define NO_CBLAS
 
 namespace Ariadne { namespace Base {
-template<typename T, unsigned int N=0> class array;
+template<typename T, size_t N=0> class array;
 template<typename T> class sequence;
   
 class BinaryWord;
